Fixed step() bug
If we break at a single byte opcode (like jump), we will continue
looping on that opcode when trying to step(). That's because each
time I will think a breakpoint had been hit.

Super hacky solution is to hold the state in the breakpoint struct.
TODO .. see if you can come up with a better solution.

Seems that that SIGINFO doesn't give me any extra information, sadly.
The signal number is the same (Trace/breakpoint trap == 5). I wonder
how I'm supposed to tell if I was hit by a breakpoint or that I just
took a new step